refactoring error handler

+++ internal/model/domain/role_entity.go
@@ -1,6 +1,6 @@
 package domain
 
 type Role struct {
-	ID   int    `gorm:"primary_key; auto_increment"`
+	ID   int64  `gorm:"primary_key; auto_increment"`
 	Name string `gorm:"size:100; not null; type:varchar(100);uniqueIndex"`
 }
